requirements.txt
algofi-py-sdk==1.0.6
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.1
msgpack==1.0.4
py-algorand-sdk==1.15.0
pycparser==2.21
pycryptodomex==3.15.0
PyNaCl==1.5.0
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.2
tzdata==2022.1
tzlocal==4.2
